procd: fix do_install task failure
The do_install task fails with the below error:
-- snip --
| install: failed to access ‘.../work/i586-poky-linux/procd/git-r0/image/lib/preinit’: No such file or directory
| WARNING: .../work/i586-poky-linux/procd/git-r0/temp/run.do_install.29662:1 exit 1 from 'install -Dm 0644 -t .../work/i586-poky-linux/procd/git-r0/image/lib/preinit .../work/i586-poky-linux/procd/git-r0/git/openwrt/package/base-files/files/lib/preinit/*'
-- snip --

as ${D}${base_libdir}/prelink directory is not created.
The "-D" option of install command, skips creating 'prelink'
directory, due to which a failure occurs as the files are unable to copy
into non-existent directory. This is same to ${D}${sysconfdir}/rc.button.
Hence, create actual directories before copying the files.

Signed-off-by: Jagadeesh Krishnanjanappa <jkrishnanjanappa@mvista.com>